Export C API's in JNI libraries
These are useful for other languages which need the FFI, so the JNI libs should have the C API's exported as well

With the split setup libraries, there is no easy way to export the C function automagically, so a manual file is used again.

This file is only needed with the JNI library.