Clean up drive class implementations
Left and right sides of the differential drive and mecanum drive classes
are now both forward and the "inversion multiplier" was removed.

The deadband functionality was moved to GenericHID::GetRawAxis() since
the deadband is generally applied to joystick axes.

MotorSafety was removed since that's handled by the drivetrain's motor
controllers already.