Add typesafe matrix wrapper over EJML to wpiutil
As a consequence of Java's limited generics system, matrix dimension
needs to be encoded using a class. Classes are generated for numbers
from 0 to 20 using GenericNumber.java.in, and can be accessed using
corresponding methods in Nat.

The primary use for this will be as part of the wpilibj implementation
of state space control