Fix PIDControllerRunner member destruction order
The mutexes in PIDControllerRunner are declared after the Notifier, and
we don't stop the Notifier in `~PIDControllerRunner()`; we call
`PIDControllerRunner::Disable()`, but that doesn't actually stop the
Notifier. It still calls this:

```
void PIDControllerRunner::Run() {
  // Ensures m_enabled check and m_controllerOutput() call occur atomically
  std::scoped_lock outputLock(m_outputMutex);
  std::unique_lock mainLock(m_thisMutex);
  if (m_enabled) {
    // Don't block other PIDControllerRunner operations on output
    mainLock.unlock();

    m_controllerOutput(m_controller.Calculate(m_measurementSource()));
  }
}
```

When we call `PIDControllerRunner::Disable()`, the Notifier is still
running, locking the mutexes, and accessing `m_enabled`. When the
PIDControllerRunner object is destructed, the PIDControllerRunner
destructor runs, then the member object destructors are called in the
reverse order in which they are declared. The mutexes are destructed
first, then the Notifier destructor is called which stops the Notifier.
There's a window between those destructor calls during which the
Notifier can run the callable and attempt to lock a mutex that no longer
exists.

To fix this, we can either call `Notifier::Disable()` in
`~PIDControllerRunner()` so the Notifier won't be running during that
destruction window, or we can declare the Notifier after all the
variables its callable uses. I opted for the latter since it's
RAII-correct.