Fix array out of bounds exception caused by parallel race group
The current index would be set to -1 by the execute method of parallelracegroup, and then an index out of bounds exception would be thrown by the end() method of sequentialcommandgroup. This change bound checks the current command index as well as only calls end at the end of parallel race group rather than during execute.