React more gracefully to attempted websocket connections
Add a branch to the server loop of ntcore to check whether a client is
attempting to connect over websockets, and if so, respond with a static
HTTP response and terminate the connection. Doing this may provide more
information to the client if its websocket implementation propogates the
HTTP error.

This also introduces a new function to NetworkStream: peek. This is used
to ensure that in the case of a valid client, handshake bytes are not
consumed by the check. The only concrete implementation of this function
that was introduced is in TCPStream.