Allow for multiple clients to the HALSim websocket server
Instead of a single weak_ptr, store connected clients as a StringMap,
where the string is an identifier determined from the connection url:
/wpilibws/<id>. The identifiers ensure that only one type of client is
connected to the simulator at a time. Clients that attempt to use an id
that is already in use receive an HTTP 409 Conflict response to the
initial handshake.