Completely remove m_hws, handle conflicts gracefully
HALSim server now checks whether an id is already used prior to the
upgrade. If the id is taken it sends HTTP 409 Conflict as a response.

All instances of m_hws have been removed and replaced by uses of
m_conns. Signature of CloseWebsocket changed to reflect this by taking
the original request URL that the identifier can be extracted from.