[wpilib] Rename SpeedController to MotorController
SpeedController and SpeedControllerGroup are retained for backwards
compatibility.  Both are deprecated and so give warnings in user code.
The MotorController interface is derived from the SpeedController interface
so that code such as SpeedController x = new VictorSP(1) continues to
compile (just with a warning).

SpeedControllerGroup and MotorControllerGroup are independent classes that
implement SpeedController and MotorController respectively.