Reset m_positionError when PIDController.reset() is called.
In addition to m_prevError and m_totalError, m_positionError needs to be reset
to 0 when reset() is called. Otherwise, the next time calculate() is
called, the old value of m_positionError will be used as the previous
error, but this is inaccurate since the caller wanted to reset the state
of the PID controller.