Adds CAN Stream API to Java through JNI bindings
The CAN Stream API allows defining an buffer to receive an
arbitrary set of CAN messages, based on an ID and a mask. Messages
are added to this queue separate of other CAN APIs. This means the
messages can be receive without impacting other APIs such as
vendor APIs.

This would enable things like detection of what devices are on the
bus, or custom decoding, without using vendor APIs.