[build] Fix CMake system library opt-ins
I'm working on setting up an AUR package for wpimath, and that platform
generally builds against system versions of libraries instead of
packaging its own. I had to make a few fixes.

-DUSE_VCPKG_EIGEN now only removes include paths for Eigen instead of
drake as well.

The USE_VCPKG flags were renamed to USE_SYSTEM since they seem general
enough for that to work (the find_package() commands work the same way
on Arch). The unofficial-libuv CMake library file only exists on vcpkg,
so we can either keep it as-is (only works on vcpkg) or provide our own
library file for finding the system install.